+++ admin-service/src/main/java/org/example/admin/service/IBorrowService.java
@@ -2,7 +2,7 @@ package org.example.admin.service;
 
 import com.baomidou.mybatisplus.extension.service.IService;
 import org.example.admin.pojo.dto.ReturnRegisterDTO;
-import org.example.admin.pojo.entity.Borrow;
+import org.example.admin.entity.Borrow;
 import org.example.admin.pojo.query.PageQuery;
 import org.example.common.result.PageResult;
 
@@ -13,16 +13,16 @@ import java.util.List;
  * 借阅记录表 服务类
  * </p>
  *
- * @author wabbybabbo
+ * @author zhengjunpeng
  * @since 2024-04-07
  */
 public interface IBorrowService extends IService<Borrow> {
 
     /**
-     * 分页查询图书
+     * 分页查询书籍信息
      *
      * @param pageQuery 分页查询时传递的数据模型
-     * @return 图书列表
+     * @return 书籍信息列表
      */
     PageResult<Borrow> pageQuery(PageQuery pageQuery);
 
@@ -32,19 +32,20 @@ public interface IBorrowService extends IService<Borrow> {
      * @param id 用户ID
      * @return 用户的借阅记录列表
      */
-    List<Borrow> getBorrows(Integer id);
+    List<Borrow> getBorrows(Long id);
 
     /**
      * 归还登记
      *
-     * @param returnRegisterDTO 用户归还借阅的图书时传递的数据模型
+     * @param returnRegisterDTO 用户归还借阅的书籍时传递的数据模型
      */
     void returnRegister(ReturnRegisterDTO returnRegisterDTO);
 
     /**
      * 借阅登记
      *
-     * @param id 图书借阅记录ID
+     * @param id 书籍借阅记录ID
      */
-    void borrowRegister(Integer id);
+    void borrowRegister(Long id);
+
 }
