+++ admin-service/src/main/java/org/example/admin/service/ICategoryService.java
@@ -3,7 +3,7 @@ package org.example.admin.service;
 import com.baomidou.mybatisplus.extension.service.IService;
 import org.example.admin.pojo.dto.CreateCategoryDTO;
 import org.example.admin.pojo.dto.UpdateCategoryDTO;
-import org.example.admin.pojo.entity.Category;
+import org.example.admin.entity.Category;
 import org.example.admin.pojo.query.PageQuery;
 import org.example.common.result.PageResult;
 
@@ -11,53 +11,54 @@ import java.util.List;
 
 /**
  * <p>
- * 图书类别表 服务类
+ * 书籍类别表 服务类
  * </p>
  *
- * @author wabbybabbo
+ * @author zhengjunpeng
  * @since 2024-04-07
  */
 public interface ICategoryService extends IService<Category> {
     /**
-     * 新建图书类别
+     * 新建书籍类别
      *
-     * @param createCategoryDTO 新建图书类别时传递的数据模型
+     * @param createCategoryDTO 新建书籍类别时传递的数据模型
      */
     void createCategory(CreateCategoryDTO createCategoryDTO);
 
     /**
-     * 查询所有图书类别项目
+     * 查询所有书籍类别
      *
-     * @return 图书类别项目列表
+     * @return 书籍类别列表
      */
     List<Category> getCategories();
 
     /**
-     * 更改图书类别项目信息
+     * 更改书籍类别信息
      *
-     * @param updateCategoryDTO 更改图书类别项目信息时传递的数据模型
+     * @param updateCategoryDTO 更改书籍类别信息时传递的数据模型
      */
     void updateCategory(UpdateCategoryDTO updateCategoryDTO);
 
     /**
-     * 删除图书类别项目
+     * 删除书籍类别
      *
-     * @param id 图书类别ID
+     * @param id 书籍类别ID
      */
-    void deleteCategory(Integer id);
+    void deleteCategory(Long id);
 
     /**
-     * 批量删除图书类别
+     * 批量删除书籍类别
      *
-     * @param ids 图书类别ID列表
+     * @param ids 书籍类别ID列表
      */
-    void batchDeleteCategories(List<Integer> ids);
+    void batchDeleteCategories(List<Long> ids);
 
     /**
-     * 分页查询图书类别
+     * 分页查询书籍类别
      *
      * @param pageQuery 分页查询时传递的数据模型
-     * @return 图书类别列表
+     * @return 书籍类别列表
      */
     PageResult<Category> pageQuery(PageQuery pageQuery);
+
 }
