+++ client-service/src/main/java/org/example/client/context/UserContext.java
@@ -1,28 +1,30 @@
 package org.example.client.context;
 
 public class UserContext {
-    private static final ThreadLocal<Integer> tl = new ThreadLocal<>();
+
+    private static final ThreadLocal<Long> tl = new ThreadLocal<>();
 
     /**
-     * 保存当前登录用户id到ThreadLocal
+     * 保存当前登录用户id到ThreadLocal中
      *
      * @param userId 用户id
      */
-    public static void setUserId(int userId) {
+    public static void setUserId(long userId) {
         tl.set(userId);
     }
 
     /**
-     * 获取当前登录用户id
+     * 从ThreadLocal中获取当前登录的用户id
      */
-    public static int getUserId() {
+    public static long getUserId() {
         return tl.get();
     }
 
     /**
-     * 移除当前登录用户id
+     * 移除ThreadLocal中当前登录的用户id
      */
     public static void removeUserId() {
         tl.remove();
     }
+
 }
